decouple zip and file verification operation
