+++ main.js
@@ -2,14 +2,13 @@
 import cors from "cors";
 import { app, BrowserWindow, dialog, ipcMain, shell } from "electron";
 import applog from "electron-log";
-import JS7z from "js7z-tools";
 import { Buffer } from "node:buffer";
-import crypto from "node:crypto";
 import fs from "node:fs";
 import * as fsPromises from "node:fs/promises";
 import path from "node:path";
 import process from "node:process";
 import { fileURLToPath } from "node:url";
+
 import { createSplashWindow, createWindow } from "./electron/createWindow.js";
 import { expressApp } from "./electron/expressServer.js";
 import {
@@ -18,6 +17,7 @@ import {
     getUserSettingsPath,
     readUserSettings,
 } from "./electron/filePath.js";
+import { fileVerification, verifyAndExtractIPC } from "./electron/zipOperation.js";
 
 const DEFAULT_PORT = 8998;
 
@@ -263,7 +263,11 @@ ipcMain.handle("check-recording-exists", async (event, key) => {
 
 // Handle playing a recording (this can be improved for streaming)
 ipcMain.handle("play-recording", async (event, key) => {
-    const filePath = path.join(await getSaveFolder(readUserSettings), "saved_recordings", `${key}.wav`);
+    const filePath = path.join(
+        await getSaveFolder(readUserSettings),
+        "saved_recordings",
+        `${key}.wav`
+    );
 
     // Check if the file exists
     try {
@@ -331,17 +335,6 @@ ipcMain.handle("open-log-folder", async () => {
     return currentLogFolder; // Send the path back to the renderer
 });
 
-// Function to calculate the SHA-256 hash of a file
-const calculateFileHash = (filePath) => {
-    return new Promise((resolve, reject) => {
-        const hash = crypto.createHash("sha256");
-        const stream = fs.createReadStream(filePath);
-        stream.on("data", (data) => hash.update(data));
-        stream.on("end", () => resolve(hash.digest("hex")));
-        stream.on("error", (err) => reject(err));
-    });
-};
-
 // Check video extracted folder
 ipcMain.handle("check-extracted-folder", async (event, folderName, zipContents) => {
     const saveFolder = await getSaveFolder(readUserSettings);
@@ -365,230 +358,7 @@ ipcMain.handle("check-extracted-folder", async (event, folderName, zipContents) 
     }
 });
 
-async function fileVerification(event, zipContents, extractedFolder) {
-    // Verify existing extracted files
-    const totalFiles = zipContents[0].extractedFiles.length;
-    let filesProcessed = 0;
-    let fileErrors = [];
-
-    for (const file of zipContents[0].extractedFiles) {
-        const extractedFilePath = path.join(extractedFolder, file.name);
-        try {
-            await fsPromises.access(extractedFilePath);
-        } catch {
-            console.log(`File missing: ${file.name}`);
-            applog.log(`File missing: ${file.name}`);
-            fileErrors.push({
-                type: "missing",
-                name: file.name,
-                message: `File missing: ${file.name}. Make sure you do not rename or accidentally delete it.`,
-            });
-            continue;
-        }
-
-        const extractedFileHash = await calculateFileHash(extractedFilePath);
-        if (extractedFileHash !== file.hash) {
-            console.log(`Hash mismatch for file: ${file.name}`);
-            applog.log(`Hash mismatch for file: ${file.name}`);
-            fileErrors.push({
-                type: "hash-mismatch",
-                name: file.name,
-                message: `Hash mismatch for file: ${file.name}. It seems like the file was either corrupted or tampered.`,
-            });
-            continue;
-        }
-
-        filesProcessed++;
-        const progressPercentage = Math.floor((filesProcessed / totalFiles) * 100);
-        event.sender.send("progress-update", progressPercentage); // Send progress update
-        applog.log("Verifying extracted file:", file.name);
-    }
-
-    if (fileErrors.length === 0) {
-        // If no missing or corrupted files, finish the verification
-        event.sender.send("verification-success", {
-            messageKey:
-                "settingPage.videoDownloadSettings.electronVerifyMessage.extractedSuccessMsg",
-            param: extractedFolder,
-        });
-        applog.log(`All extracted files are verified for "${extractedFolder}"`);
-        return;
-    } else {
-        // Send all errors as an array
-        event.sender.send("verification-errors", fileErrors);
-        applog.log(
-            `Verification found errors in extracted files for "${extractedFolder}"`,
-            fileErrors
-        );
-        return;
-    }
-}
-
-// Handle verify and extract process using JS7z
-ipcMain.on("verify-and-extract", async (event, zipFileData) => {
-    const { zipFile, zipHash, zipContents } = zipFileData;
-    const saveFolder = await getSaveFolder(readUserSettings);
-    const videoFolder = path.join(saveFolder, "video_files");
-    const zipFilePath = path.join(videoFolder, zipFile);
-    const extractedFolder = path.join(videoFolder, zipFile.replace(".7z", ""));
-
-    // Step 0: Check if ZIP file exists first
-    let zipExists = false;
-    try {
-        await fsPromises.access(zipFilePath);
-        zipExists = true;
-    } catch {
-        zipExists = false;
-    }
-
-    if (zipExists) {
-        // ZIP exists: proceed with hash verification and extraction
-        console.log(`Starting verification for ${zipFile}`);
-        applog.log(`Starting verification for ${zipFile}`);
-        try {
-            const js7z = await JS7z({
-                print: (text) => {
-                    console.log(`7-Zip output: ${text}`);
-                    applog.log(`7-Zip output: ${text}`);
-                    if (text.includes("%")) {
-                        const match = text.match(/\s+(\d+)%/); // Extract percentage from output
-                        if (match) {
-                            const percentage = parseInt(match[1], 10);
-                            event.sender.send("progress-update", percentage);
-                        }
-                    }
-                },
-                printErr: (errText) => {
-                    console.error(`7-Zip error: ${errText}`);
-                    applog.error(`7-Zip error: ${errText}`);
-                    event.sender.send("verification-error", `7-Zip error: ${errText}`);
-                },
-                onAbort: (reason) => {
-                    console.error(`7-Zip aborted: ${reason}`);
-                    applog.error(`7-Zip aborted: ${reason}`);
-                    event.sender.send("verification-error", `7-Zip aborted: ${reason}`);
-                },
-                onExit: (exitCode) => {
-                    if (exitCode === 0) {
-                        console.log(`7-Zip exited successfully with code ${exitCode}`);
-                        applog.log(`7-Zip exited successfully with code ${exitCode}`);
-                    } else {
-                        console.error(`7-Zip exited with error code: ${exitCode}`);
-                        applog.error(`7-Zip exited with error code: ${exitCode}`);
-                        event.sender.send(
-                            "verification-error",
-                            `7-Zip exited with error code: ${exitCode}`
-                        );
-                    }
-                },
-            });
-
-            // Mount the local file system to the JS7z instance using NODEFS
-            js7z.FS.mkdir("/mnt");
-            js7z.FS.mount(js7z.NODEFS, { root: videoFolder }, "/mnt");
-
-            const emZipFilePath = `/mnt/${zipFile}`;
-            const emExtractedFolder = `/mnt/${zipFile.replace(".7z", "")}`;
-
-            // Step 1: Verifying ZIP file hash
-            event.sender.send(
-                "progress-text",
-                "settingPage.videoDownloadSettings.electronVerifyMessage.zipFileVerifying"
-            );
-            const fileHash = await calculateFileHash(zipFilePath);
-            if (fileHash !== zipHash) {
-                applog.error(
-                    `Hash mismatch for ${zipFile}. It seems like the zip file was either corrupted or tampered.`
-                );
-                event.sender.send("verification-error", {
-                    messageKey:
-                        "settingPage.videoDownloadSettings.electronVerifyMessage.zipHashMismatchMsg",
-                    param: zipFile,
-                });
-                return;
-            }
-            event.sender.send("progress-text", "ZIP file verified");
-
-            // Step 2: Extracting ZIP file
-            event.sender.send(
-                "progress-text",
-                "settingPage.videoDownloadSettings.electronVerifyMessage.zipExtractingMsg"
-            );
-            js7z.callMain(["x", emZipFilePath, `-o${emExtractedFolder}`]);
-
-            js7z.onExit = async (exitCode) => {
-                if (exitCode !== 0) {
-                    applog.error(`Error extracting ${zipFile}`);
-                    event.sender.send("verification-error", {
-                        messageKey:
-                            "settingPage.videoDownloadSettings.electronVerifyMessage.zipErrorMsg",
-                        param: zipFile,
-                    });
-                    return;
-                }
-
-                console.log(`Extraction successful for ${zipFile}`);
-                applog.log(`Extraction successful for ${zipFile}`);
-
-                // Step 3: Verifying extracted files
-                event.sender.send(
-                    "progress-text",
-                    "settingPage.videoDownloadSettings.verifyinProgressMsg"
-                );
-                await fileVerification(event, zipContents, extractedFolder);
-
-                // Clean up the ZIP file after successful extraction and verification
-                try {
-                    await fsPromises.unlink(zipFilePath);
-                    console.log(`Deleted ZIP file: ${zipFilePath}`);
-                    applog.log(`Extraction successful for ${zipFile}`);
-                } catch (err) {
-                    console.error(`Failed to delete ZIP file: ${err.message}`);
-                    applog.error(`Failed to delete ZIP file: ${err.message}`);
-                }
-
-                event.sender.send("verification-success", {
-                    messageKey:
-                        "settingPage.videoDownloadSettings.electronVerifyMessage.zipSuccessMsg",
-                    param: zipFile,
-                });
-                applog.log(`Successfully verified and extracted ${zipFile}`);
-            };
-        } catch (err) {
-            console.error(`Error processing ${zipFile}: ${err.message}`);
-            event.sender.send("verification-error", {
-                messageKey: "settingPage.videoDownloadSettings.electronVerifyMessage.zipErrorMsg",
-                param: zipFile,
-                errorMessage: err.message,
-            });
-        }
-    } else {
-        // ZIP does not exist: check for extracted folder and verify its contents
-        console.log(
-            `ZIP file does not exist: ${zipFilePath}. It could be extracted before. Proceeding with extracted folder verification...`
-        );
-        applog.log(
-            `ZIP file does not exist: ${zipFilePath}. It could be extracted before. Proceeding with extracted folder verification...`
-        );
-        try {
-            await fsPromises.access(extractedFolder);
-            console.log(`Extracted folder already exists: ${extractedFolder}`);
-            applog.log(`Extracted folder already exists: ${extractedFolder}`);
-            event.sender.send(
-                "progress-text",
-                "settingPage.videoDownloadSettings.verifyinProgressMsg"
-            );
-            await fileVerification(event, zipContents, extractedFolder);
-        } catch {
-            console.log(`Extracted folder does not exist: ${extractedFolder}`);
-            applog.log(`Extracted folder does not exist: ${extractedFolder}`);
-            event.sender.send("verification-error", {
-                messageKey: "settingPage.videoDownloadSettings.verifyFailedMessage",
-                param: extractedFolder,
-            });
-        }
-    }
-});
+verifyAndExtractIPC(() => getSaveFolder(readUserSettings), fileVerification);
 
 // Listen for logging messages from the renderer process
 ipcMain.on("renderer-log", (event, logMessage) => {
